fixed putadrr

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: fbicane <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 16:59:43 by fbicane           #+#    #+#             */
-/*   Updated: 2024/11/25 12:51:58 by fbicane          ###   ########.fr       */
+/*   Updated: 2024/11/26 13:09:59 by fbicane          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 int	ft_putchar(char c);
 int	ft_putstr(char *s);
 int	ft_putnbr(int n);
-int	ft_puthexa(unsigned int n, char b);
+int	ft_puthexa(unsigned long n, char b);
 int	ft_nb_count(long nb);
 int	ft_putunint(unsigned int n);
 int	ft_putadrr(void *ptr);
